requirements.txt

Django==4.1.1
django-nested-admin==3.4.0
ocrmypdf==13.7.0
pdfplumber==0.7.4
psycopg2==2.9.3
pytesseract==0.3.10

